Factor out ScanLauncher from create view
