+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/create/CreateView.kt
@@ -4,12 +4,11 @@ import android.annotation.SuppressLint
 import android.app.DatePickerDialog
 import android.app.TimePickerDialog
 import android.content.ActivityNotFoundException
+import android.content.Context
 import android.content.Intent
 import android.location.Location
 import android.net.Uri
 import android.widget.Toast
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
 import androidx.compose.animation.shrinkVertically
@@ -68,6 +67,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalResources
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.dp
@@ -89,6 +89,7 @@ import nz.eloque.foss_wallet.model.PassRelevantDate
 import nz.eloque.foss_wallet.model.PassType
 import nz.eloque.foss_wallet.ui.Screen
 import nz.eloque.foss_wallet.ui.components.ImagePicker
+import nz.eloque.foss_wallet.ui.screens.scan.ScanLauncher
 import nz.eloque.foss_wallet.ui.screens.settings.ComboBox
 import java.time.ZoneId
 import java.time.ZonedDateTime
@@ -174,44 +175,22 @@ fun CreateView(
     val allColorsBlank = backgroundColor == null && foregroundColor == null && labelColor == null
 
     val scanLauncher =
-        rememberLauncherForActivityResult(
-            contract = ActivityResultContracts.StartActivityForResult(),
-            onResult = { activityResult ->
-                if (activityResult.resultCode == android.app.Activity.RESULT_OK) {
-                    val resultData = activityResult.data
-                    val contents = resultData?.getStringExtra(ScanActivity.EXTRA_RESULT)
-                    if (contents != null) {
-                        val formatName = resultData.getStringExtra(ScanActivity.EXTRA_RESULT_FORMAT)
-                        if (activeBarcodeIndex !in barcodes.indices) return@rememberLauncherForActivityResult
-                        barcodes =
-                            barcodes.mapIndexed { index, barcode ->
-                                if (index != activeBarcodeIndex) {
-                                    barcode
-                                } else {
-                                    val scannedFormat =
-                                        try {
-                                            BarcodeFormat.valueOf(formatName ?: BarcodeFormat.QR_CODE.name)
-                                        } catch (_: IllegalArgumentException) {
-                                            Toast
-                                                .makeText(
-                                                    context,
-                                                    resources.getString(R.string.no_barcode_format_given),
-                                                    Toast.LENGTH_SHORT,
-                                                ).show()
-                                            BarcodeFormat.QR_CODE
-                                        }
-                                    barcode.copy(
-                                        message = contents,
-                                        altText = contents,
-                                        format = scannedFormat,
-                                    )
-                                }
-                            }
-                        detailsExpanded = true
+        ScanLauncher.launch {
+            if (activeBarcodeIndex !in barcodes.indices) return@launch
+            barcodes =
+                barcodes.mapIndexed { index, barcode ->
+                    if (index != activeBarcodeIndex) {
+                        barcode
+                    } else {
+                        barcode.copy(
+                            message = it.message,
+                            altText = it.altText ?: it.message,
+                            format = it.format,
+                        )
                     }
                 }
-            },
-        )
+            detailsExpanded = true
+        }
 
     LaunchedEffect(startMode, initialScanHandled) {
         if (startMode == CreateStartMode.Scan && !initialScanHandled) {
@@ -723,8 +702,7 @@ private fun ColorPickerRow(
                 readOnly = true,
                 label = { Text(label) },
                 textStyle =
-                    androidx.compose.ui.text
-                        .TextStyle(fontFamily = FontFamily.Monospace),
+                    TextStyle(fontFamily = FontFamily.Monospace),
                 leadingIcon = {
                     Row(
                         modifier = Modifier.padding(start = 10.dp),
@@ -942,7 +920,7 @@ private fun LocationPickerDialog(
 }
 
 private fun openDateTimePicker(
-    context: android.content.Context,
+    context: Context,
     initial: ZonedDateTime?,
     onPicked: (ZonedDateTime) -> Unit,
 ) {
